requirements.txt
# Core dependencies
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# CSV processing
pandas>=2.2.0

# Database (PostgreSQL - solo para producción)
# psycopg2-binary>=2.9.9  # Descomentar si usas PostgreSQL

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Code quality
flake8>=6.1.0
black>=23.12.0

# Utilities
python-dateutil>=2.8.2
